# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: Parcel Intelligence API
  version: 0.1.0
paths:
  /api/v1/auth/register:
    post:
      tags:
      - auth
      summary: Register
      description: Register a new user account.
      operationId: register_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: Login
      description: Authenticate with email and password to receive JWT tokens.
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/refresh:
    post:
      tags:
      - auth
      summary: Refresh Token Endpoint
      description: Obtain a new access token using a valid refresh token.
      operationId: refresh_token_endpoint_api_v1_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/me:
    get:
      tags:
      - auth
      summary: Get Current User Profile
      description: Retrieve profile of the currently authenticated user.
      operationId: get_current_user_profile_api_v1_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
      security:
      - HTTPBearer: []
...
  /api/v1/consents:
    post:
      tags:
      - consents
      summary: Request Consent
      description: Submit a DPI Consent request to inspect sensitive parcel records.
      operationId: request_consent_api_v1_consents_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentCreateRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/consents/{consent_id}:
    patch:
      tags:
      - consents
      summary: Handle Consent Action
      description: Approve or revoke a pending/active consent grant.
      operationId: handle_consent_action_api_v1_consents__consent_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: consent_id
        in: path
        required: true
        schema:
          type: integer
          title: Consent Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentActionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/consents/mine:
    get:
      tags:
      - consents
      summary: Get My Consents
      description: List incoming and outgoing consent grants relevant to the current
        user.
      operationId: get_my_consents_api_v1_consents_mine_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ConsentResponse'
                type: array
                title: Response Get My Consents Api V1 Consents Mine Get
      security:
      - HTTPBearer: []
  /api/v1/alerts:
    get:
      tags:
      - alerts
      summary: List Alerts
      description: List cadastral boundary shift and land use change alerts.
      operationId: list_alerts_api_v1_alerts_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlertResponse'
                title: Response List Alerts Api V1 Alerts Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/alerts/{alert_id}:
    patch:
      tags:
      - alerts
      summary: Update Alert Status
      description: Update change detection alert status (restricted to officers/administrators).
      operationId: update_alert_status_api_v1_alerts__alert_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: alert_id
        in: path
        required: true
        schema:
          type: integer
          title: Alert Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertStatusUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AlertResponse:
      properties:
        id:
          type: integer
          title: Id
        parcel_id:
          type: integer
          title: Parcel Id
        ulpin:
          type: string
          title: Ulpin
        alert_type:
          type: string
          title: Alert Type
        confidence_score:
          type: number
          title: Confidence Score
        source:
          type: string
          title: Source
        status:
          type: string
          title: Status
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - parcel_id
      - ulpin
      - alert_type
      - confidence_score
      - source
      - status
      - created_at
      title: AlertResponse
    AlertStatusUpdate:
      properties:
        status:
          type: string
          pattern: ^(unresolved|investigating|dismissed|resolved)$
          title: Status
      type: object
      required:
      - status
      title: AlertStatusUpdate
...
    ConsentActionRequest:
      properties:
        action:
          type: string
          pattern: ^(approve|revoke)$
          title: Action
          description: '''approve'' or ''revoke'''
        valid_days:
          anyOf:
          - type: integer
            maximum: 365.0
            minimum: 1.0
          - type: 'null'
          title: Valid Days
          default: 30
      type: object
      required:
      - action
      title: ConsentActionRequest
    ConsentCreateRequest:
      properties:
        ulpin:
          type: string
          title: Ulpin
        purpose:
          type: string
          maxLength: 255
          minLength: 3
          title: Purpose
          description: e.g. loan_due_diligence, title_verification
        scope:
          items:
            type: string
          type: array
          title: Scope
          description: 'List of scopes: ror, encumbrance, tax, mutations'
          default:
          - ror
          - encumbrance
          - tax
      type: object
      required:
      - ulpin
      - purpose
      title: ConsentCreateRequest
    ConsentResponse:
      properties:
        id:
          type: integer
          title: Id
        requester_user_id:
          type: integer
          title: Requester User Id
        requester_name:
          type: string
          title: Requester Name
        parcel_id:
          type: integer
          title: Parcel Id
        ulpin:
          type: string
          title: Ulpin
        purpose:
          type: string
          title: Purpose
        scope:
          items:
            type: string
          type: array
          title: Scope
        status:
          type: string
          title: Status
        valid_until:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Valid Until
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - requester_user_id
      - requester_name
      - parcel_id
      - ulpin
      - purpose
      - scope
      - status
      - created_at
      title: ConsentResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TokenRefreshRequest:
      properties:
        refresh_token:
          type: string
          title: Refresh Token
      type: object
      required:
      - refresh_token
      title: TokenRefreshRequest
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
        user_id:
          type: integer
          title: User Id
        role:
          type: string
          title: Role
        name:
          type: string
          title: Name
      type: object
      required:
      - access_token
      - refresh_token
      - expires_in
      - user_id
      - role
      - name
      title: TokenResponse
    UserLogin:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: UserLogin
    UserRegister:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        email:
          type: string
          maxLength: 255
          minLength: 3
          title: Email
        password:
          type: string
          maxLength: 100
          minLength: 6
          title: Password
        role:
          type: string
          title: Role
          default: citizen
        mobile:
          anyOf:
          - type: string
          - type: 'null'
          title: Mobile
        state_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: State Id
      type: object
      required:
      - name
      - email
      - password
      title: UserRegister
    UserResponse:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        email:
          type: string
          title: Email
        role:
          type: string
          title: Role
        mobile:
          anyOf:
          - type: string
          - type: 'null'
          title: Mobile
        state_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: State Id
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
      type: object
      required:
      - id
      - name
      - email
      - role
      title: UserResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer